refactor: separate lotto creation in winning lotto
